refactor : sql Image 타입 BIGINT, java image 타입 Long으로 변경

+++ src/main/java/cart/product/dao/ProductDao.java
@@ -1,17 +1,15 @@
 package cart.product.dao;
 
-import java.sql.PreparedStatement;
-import java.util.List;
-
+import cart.product.entity.Product;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
-import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
-import cart.product.entity.Product;
+import java.sql.PreparedStatement;
+import java.util.List;
 
 @Repository
 public class ProductDao {
@@ -22,7 +20,7 @@ public class ProductDao {
 			resultSet.getLong("id"),
 			resultSet.getString("name"),
 			resultSet.getString("image"),
-			resultSet.getInt("price"),
+			resultSet.getLong("price"),
 			resultSet.getTimestamp("created_at").toLocalDateTime(),
 			resultSet.getTimestamp("updated_at").toLocalDateTime()
 		);
@@ -48,7 +46,7 @@ public class ProductDao {
 		jdbcTemplate.update(connection -> {
 			PreparedStatement ps = connection.prepareStatement(sql, new String[] {"id"});
 			ps.setString(1, product.getName());
-			ps.setInt(2, product.getPrice());
+			ps.setLong(2, product.getPrice());
 			ps.setString(3, product.getImage());
 			return ps;
 		}, keyHolder);
