+++ src/test/java/cart/product/dao/ProductDaoTest.java
@@ -28,7 +28,7 @@ class ProductDaoTest {
 		rs.getLong("id"),
 		rs.getString("name"),
 		rs.getString("image"),
-		rs.getInt("price"),
+		rs.getLong("price"),
 		rs.getTimestamp("created_at").toLocalDateTime(),
 		rs.getTimestamp("updated_at").toLocalDateTime()
 	);
@@ -36,7 +36,7 @@ class ProductDaoTest {
 	@BeforeEach
 	public void setProduct() {
 		productDao = new ProductDao(jdbcTemplate);
-		product = new Product(null, "name", "image", 10000, null, null);
+		product = new Product(null, "name", "image", 10000L, null, null);
 	}
 
 	@Test
@@ -68,7 +68,7 @@ class ProductDaoTest {
 		long id = productDao.save(product);
 
 		// when
-		Product newProduct = new Product(null, "newName", "newImage", 0, null, null);
+		Product newProduct = new Product(null, "newName", "newImage", 0L, null, null);
 		productDao.updateById(id, newProduct);
 
 		// then
